20211021: as runnable app with controller, avro, assembly
