+++ spring-boot-kafka-admin/src/assembly/assembly.xml
@@ -1,0 +1,46 @@
+<assembly
+        xmlns="http://maven.apache.org/ASSEMBLY/2.0.0"
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:schemaLocation="http://maven.apache.org/ASSEMBLY/2.0.0 http://maven.apache.org/xsd/assembly-2.0.0.xsd">
+    <id>assembly</id>
+    <includeBaseDirectory>false</includeBaseDirectory>
+    <baseDirectory>${project.artifactId}</baseDirectory>
+    <formats>
+        <format>jar</format>
+        <format>zip</format>
+        <format>tar.gz</format>
+    </formats>
+    <dependencySets>
+        <dependencySet>
+            <excludes>
+                <exclude>commons-logging:commons-logging</exclude>
+                <exclude>log4j:log4j</exclude>
+            </excludes>
+        </dependencySet>
+    </dependencySets>
+    <fileSets>
+        <fileSet>
+            <directory>${basedir}/src/config</directory>
+            <outputDirectory>conf</outputDirectory>
+            <includes>
+                <include>**/*.*</include>
+            </includes>
+            <excludes>
+                <exclude>.git</exclude>
+                <exclude>.svn</exclude>
+            </excludes>
+        </fileSet>
+    </fileSets>
+    <files>
+        <file>
+            <source>${basedir}/src/main/resources/AWESOME.txt</source>
+            <outputDirectory>/MyFiles</outputDirectory>
+            <filtered>true</filtered>
+        </file>
+        <file>
+            <source>${basedir}/src/main/resources/LICENSE.txt</source>
+            <outputDirectory>/License</outputDirectory>
+            <filtered>true</filtered>
+        </file>
+    </files>
+</assembly>
