+++ spring-boot-kafka-admin/src/config/DevLocal/local.properties
@@ -1,0 +1,2 @@
+aaa=bbb
+
