+++ spring-boot-kafka-admin/src/main/java/com/oopsmails/springboot/kafka/admin/KafkaApplication.java
@@ -1,32 +1,20 @@
 package com.oopsmails.springboot.kafka.admin;
 
-import lombok.Data;
+import com.oopsmails.springboot.kafka.admin.config.MessageListener;
+import com.oopsmails.springboot.kafka.admin.config.MessageProducer;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.context.annotation.Bean;
-import org.springframework.kafka.annotation.KafkaListener;
-import org.springframework.kafka.annotation.TopicPartition;
-import org.springframework.kafka.core.KafkaTemplate;
-import org.springframework.kafka.support.KafkaHeaders;
-import org.springframework.kafka.support.SendResult;
-import org.springframework.messaging.handler.annotation.Header;
-import org.springframework.messaging.handler.annotation.Payload;
-import org.springframework.util.concurrent.ListenableFuture;
-import org.springframework.util.concurrent.ListenableFutureCallback;
 
-import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 
 @SpringBootApplication
 @EnableAutoConfiguration
 @Slf4j
 public class KafkaApplication {
-
     public static void main(String[] args) throws Exception {
 
         ConfigurableApplicationContext context = SpringApplication.run(KafkaApplication.class, args);
@@ -42,7 +30,7 @@ public class KafkaApplication {
          * headersKafkaListenerContainerFactory as container factory
          */
         producer.sendMessage("Hello, World!");
-        listener.latch.await(10, TimeUnit.SECONDS);
+        listener.getLatch().await(10, TimeUnit.SECONDS);
 
         /*
          * Sending message to a topic with 5 partition,
@@ -53,7 +41,7 @@ public class KafkaApplication {
         for (int i = 0; i < 5; i++) {
             producer.sendMessageToPartion("Hello To Partitioned Topic!", i);
         }
-        listener.partitionLatch.await(10, TimeUnit.SECONDS);
+        listener.getPartitionLatch().await(10, TimeUnit.SECONDS);
 
         /*
          * Sending message to 'filtered' topic. As per listener
@@ -62,7 +50,7 @@ public class KafkaApplication {
          */
         producer.sendMessageToFiltered("Hello Oopstopic!");
         producer.sendMessageToFiltered("Hello World!");
-        listener.filterLatch.await(10, TimeUnit.SECONDS);
+        listener.getFilterLatch().await(10, TimeUnit.SECONDS);
 
         /*
          * Sending message to 'greeting' topic. This will send
@@ -70,9 +58,9 @@ public class KafkaApplication {
          * greetingKafkaListenerContainerFactory.
          */
         producer.sendGreetingMessage(new Greeting("Greetings", "World!"));
-        listener.greetingLatch.await(10, TimeUnit.SECONDS);
+        listener.getGreetingLatch().await(10, TimeUnit.SECONDS);
 
-        context.close();
+//        context.close();
     }
 
     @Bean
@@ -85,111 +73,4 @@ public class KafkaApplication {
         return new MessageListener();
     }
 
-    public static class MessageProducer {
-
-        @Autowired
-        private KafkaTemplate<String, String> kafkaTemplate;
-
-        @Autowired
-        private KafkaTemplate<String, Greeting> greetingKafkaTemplate;
-
-        @Value(value = "${message.topic.name}")
-        private String topicName;
-
-        @Value(value = "${partitioned.topic.name}")
-        private String partitionedTopicName;
-
-        @Value(value = "${filtered.topic.name}")
-        private String filteredTopicName;
-
-        @Value(value = "${greeting.topic.name}")
-        private String greetingTopicName;
-
-        public void sendMessage(String message) {
-
-            ListenableFuture<SendResult<String, String>> future = kafkaTemplate.send(topicName, message);
-
-            future.addCallback(new ListenableFutureCallback<SendResult<String, String>>() {
-
-                @Override
-                public void onSuccess(SendResult<String, String> result) {
-                    System.out.println("Sent message=[" + message + "] with offset=[" + result.getRecordMetadata().offset() + "]");
-                   log.info("### -> Sent message=[" + message + "] with offset=[" + result.getRecordMetadata().offset() + "]");
-                }
-
-                @Override
-                public void onFailure(Throwable ex) {
-                    System.out.println("Unable to send message=[" + message + "] due to : " + ex.getMessage());
-                    log.info("### -> Unable to send message=[" + message + "] due to : " + ex.getMessage());
-                }
-            });
-        }
-
-        public void sendMessageToPartion(String message, int partition) {
-            kafkaTemplate.send(partitionedTopicName, partition, null, message);
-        }
-
-        public void sendMessageToFiltered(String message) {
-            kafkaTemplate.send(filteredTopicName, message);
-        }
-
-        public void sendGreetingMessage(Greeting greeting) {
-            log.info("### -> sendGreetingMessage message=[" + greeting + "] ");
-            greetingKafkaTemplate.send(greetingTopicName, greeting);
-        }
-    }
-
-    @Data
-    public static class MessageListener {
-
-        private CountDownLatch latch = new CountDownLatch(3);
-
-        private CountDownLatch partitionLatch = new CountDownLatch(2);
-
-        private CountDownLatch filterLatch = new CountDownLatch(2);
-
-        private CountDownLatch greetingLatch = new CountDownLatch(1);
-
-        @KafkaListener(topics = "${message.topic.name}", groupId = "foo", containerFactory = "fooKafkaListenerContainerFactory")
-        public void listenGroupFoo(String message) {
-            System.out.println("Received Messasge in group 'foo': " + message);
-            log.info("### -> Received Messasge in group 'foo': " + message);
-            latch.countDown();
-        }
-
-        @KafkaListener(topics = "${message.topic.name}", groupId = "bar", containerFactory = "barKafkaListenerContainerFactory")
-        public void listenGroupBar(String message) {
-            System.out.println("Received Messasge in group 'bar': " + message);
-            log.info("### -> Received Messasge in group 'bar': " + message);
-            latch.countDown();
-        }
-
-        @KafkaListener(topics = "${message.topic.name}", containerFactory = "headersKafkaListenerContainerFactory")
-        public void listenWithHeaders(@Payload String message, @Header(KafkaHeaders.RECEIVED_PARTITION_ID) int partition) {
-            System.out.println("Received Messasge: " + message + " from partition: " + partition);
-            log.info("### -> Received Messasge: " + message + " from partition: " + partition);
-            latch.countDown();
-        }
-
-        @KafkaListener(topicPartitions = @TopicPartition(topic = "${partitioned.topic.name}", partitions = {"0", "3"}))
-        public void listenToParition(@Payload String message, @Header(KafkaHeaders.RECEIVED_PARTITION_ID) int partition) {
-            System.out.println("Received Message: " + message + " from partition: " + partition);
-            log.info("### -> Received Message: " + message + " from partition: " + partition);
-            this.partitionLatch.countDown();
-        }
-
-        @KafkaListener(topics = "${filtered.topic.name}", containerFactory = "filterKafkaListenerContainerFactory")
-        public void listenWithFilter(String message) {
-            System.out.println("Recieved Message in filtered listener: " + message);
-            log.info("### -> Recieved Message in filtered listener: " + message);
-            this.filterLatch.countDown();
-        }
-
-        @KafkaListener(topics = "${greeting.topic.name}", containerFactory = "greetingKafkaListenerContainerFactory")
-        public void greetingListener(Greeting greeting) {
-            System.out.println("Recieved greeting message: " + greeting);
-            log.info("### -> Recieved greeting message: " + greeting);
-            this.greetingLatch.countDown();
-        }
-    }
 }
