+++ spring-boot-kafka-admin/src/main/java/com/oopsmails/springboot/kafka/admin/config/MessageProducer.java
@@ -1,0 +1,68 @@
+package com.oopsmails.springboot.kafka.admin.config;
+
+import com.oopsmails.springboot.kafka.admin.Greeting;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.kafka.core.KafkaTemplate;
+import org.springframework.kafka.support.SendResult;
+import org.springframework.util.concurrent.ListenableFuture;
+import org.springframework.util.concurrent.ListenableFutureCallback;
+
+@Slf4j
+public class MessageProducer {
+
+    @Autowired
+    private KafkaTemplate<String, String> kafkaTemplate;
+
+    @Autowired
+    private KafkaTemplate<String, Greeting> greetingKafkaTemplate;
+
+    @Value(value = "${message.topic.name}")
+    private String topicName;
+
+    @Value(value = "${partitioned.topic.name}")
+    private String partitionedTopicName;
+
+    @Value(value = "${filtered.topic.name}")
+    private String filteredTopicName;
+
+    @Value(value = "${greeting.topic.name}")
+    private String greetingTopicName;
+
+    @Value(value = "${person.topic.name}")
+    private String personTopicName;
+
+    public void sendMessage(String message) {
+
+        ListenableFuture<SendResult<String, String>> future = kafkaTemplate.send(topicName, message);
+
+        future.addCallback(new ListenableFutureCallback<SendResult<String, String>>() {
+
+            @Override
+            public void onSuccess(SendResult<String, String> result) {
+                System.out.println("Sent message=[" + message + "] with offset=[" + result.getRecordMetadata().offset() + "]");
+                log.info("### -> Sent message=[" + message + "] with offset=[" + result.getRecordMetadata().offset() + "]");
+            }
+
+            @Override
+            public void onFailure(Throwable ex) {
+                System.out.println("Unable to send message=[" + message + "] due to : " + ex.getMessage());
+                log.info("### -> Unable to send message=[" + message + "] due to : " + ex.getMessage());
+            }
+        });
+    }
+
+    public void sendMessageToPartion(String message, int partition) {
+        kafkaTemplate.send(partitionedTopicName, partition, null, message);
+    }
+
+    public void sendMessageToFiltered(String message) {
+        kafkaTemplate.send(filteredTopicName, message);
+    }
+
+    public void sendGreetingMessage(Greeting greeting) {
+        log.info("### -> sendGreetingMessage message=[" + greeting + "] ");
+        greetingKafkaTemplate.send(greetingTopicName, greeting);
+    }
+}
