+++ spring-boot-kafka-admin/src/main/java/com/oopsmails/springboot/kafka/admin/config/MessageProducerPerson.java
@@ -1,0 +1,4 @@
+package com.oopsmails.springboot.kafka.admin.config;
+
+public class MessageProducerPerson {
+}
