+++ spring-boot-kafka-admin/src/main/resources/AWESOME.txt
@@ -1,0 +1,1 @@
+AWESOME.txt
