+++ spring-boot-kafka-admin/src/main/resources/application.properties
@@ -1,5 +1,13 @@
+server.port=8889
+server.servlet.context-path=msg
+
 kafka.bootstrapAddress=localhost:9092
+
 message.topic.name=oopstopic
 greeting.topic.name=greeting
 filtered.topic.name=filtered
 partitioned.topic.name=partitioned
+
+person.topic.name=person_topic
+
+
