+++ spring-boot-kafka-admin/src/main/resources/avro/person.avsc
@@ -1,0 +1,17 @@
+{
+	"namespace": "com.oopsmails.avro.dto",
+	"type": "record",
+	"name": "PersonDto",
+	"fields": [
+		{
+			"name": "firstName",
+			"type": "string",
+			"doc": "the first name of a person"
+		},
+		{
+			"name": "lastName",
+			"type": "string",
+			"doc": "the last name of a person"
+		}	
+	]
+}
