+++ spring-boot-kafka-admin/src/test/resources/application.properties
@@ -3,3 +3,5 @@ message.topic.name=oopstopic
 greeting.topic.name=greeting
 filtered.topic.name=filtered
 partitioned.topic.name=partitioned
+
+person.topic.name=person_topic
