+++ spring-boot-kafka-avro/spring-boot-kafka-avro-basic/spring-boot-kafka-avro-basic-consumer/src/main/java/com/oopsmails/service/KafkaConsumerService.java
@@ -26,39 +26,39 @@ public class KafkaConsumerService {
     @Autowired
     ObjectMapper objectMapper;
 
-//    @KafkaListener(topics = "${spring.kafka.topic-name}",
-//            containerFactory = "kafkaListenerContainerFactory",
-//            id = "mainListener")
-////    @KafkaListener(topics = "${spring.kafka.topic-name}")
-//    public void listen(@Payload PersonDto personDto, @Header(KafkaHeaders.RECEIVED_PARTITION_ID) int partition) {
-//        logger.info("Receiving message from Kafka :: personDto :: {}", personDto + " from partition: " + partition);
-//        Person person = Person.builder()
-//                .firstName(personDto.getFirstName().toString())
-//                .lastName(personDto.getLastName().toString())
-//                .build();
-//        businessDomainService.postProcessReceivedMessage(person);
-//    }
-
     @KafkaListener(topics = "${spring.kafka.topic-name}",
-            containerFactory = "consumerFactory",
+            containerFactory = "kafkaListenerContainerFactory",
             id = "mainListener")
 //    @KafkaListener(topics = "${spring.kafka.topic-name}")
-    public void listenGeneric(@Payload String messageStr, @Header(KafkaHeaders.RECEIVED_PARTITION_ID) int partition) {
-        logger.info("listenGeneric, Receiving message from Kafka :: messageStr :: {}", messageStr + " from partition: " + partition);
-//        businessDomainService.postProcessReceivedMessage(person);
-
-        try {
-            PersonDto personDto = objectMapper.readValue(messageStr, PersonDto.class);
-            log.info("listenGeneric, converted to PersonDto: {}", personDto);
-            Person person = Person.builder()
-                    .firstName(personDto.getFirstName().toString())
-                    .lastName(personDto.getLastName().toString())
-                    .build();
-            log.info("listenGeneric, invoking next @Service, person: {}", personDto);
-            businessDomainService.postProcessReceivedMessage(person);
-        } catch (JsonProcessingException e) {
-            log.error("listenGeneric, cannot convert to PersonDto - cause: {}", e.getMessage());
-        }
+    public void listen(@Payload PersonDto personDto, @Header(KafkaHeaders.RECEIVED_PARTITION_ID) int partition) {
+        logger.info("Receiving message from Kafka :: personDto :: {}", personDto + " from partition: " + partition);
+        Person person = Person.builder()
+                .firstName(personDto.getFirstName().toString())
+                .lastName(personDto.getLastName().toString())
+                .build();
+        businessDomainService.postProcessReceivedMessage(person);
     }
+
+//    @KafkaListener(topics = "${spring.kafka.topic-name}",
+//            containerFactory = "consumerFactory",
+//            id = "mainListener")
+////    @KafkaListener(topics = "${spring.kafka.topic-name}")
+//    public void listenGeneric(@Payload String messageStr, @Header(KafkaHeaders.RECEIVED_PARTITION_ID) int partition) {
+//        logger.info("listenGeneric, Receiving message from Kafka :: messageStr :: {}", messageStr + " from partition: " + partition);
+////        businessDomainService.postProcessReceivedMessage(person);
+//
+//        try {
+//            PersonDto personDto = objectMapper.readValue(messageStr, PersonDto.class);
+//            log.info("listenGeneric, converted to PersonDto: {}", personDto);
+//            Person person = Person.builder()
+//                    .firstName(personDto.getFirstName().toString())
+//                    .lastName(personDto.getLastName().toString())
+//                    .build();
+//            log.info("listenGeneric, invoking next @Service, person: {}", personDto);
+//            businessDomainService.postProcessReceivedMessage(person);
+//        } catch (JsonProcessingException e) {
+//            log.error("listenGeneric, cannot convert to PersonDto - cause: {}", e.getMessage());
+//        }
+//    }
 
 }
