+++ spring-boot-kafka-docker/src/main/java/com/oopsmails/springboot/kafka/admin/config/KafkaConsumer.java
@@ -1,4 +1,4 @@
-package com.oopsmails.springboot.kafka.config;
+package com.oopsmails.springboot.kafka.admin.config;
 
 import lombok.Data;
 import lombok.extern.slf4j.Slf4j;
